requirements.txt
requests
python-dotenv
PyMuPDF